Refactoring, add mappers, lombok.

+++ src/main/java/com/pe/shopping/cart/dao/VentaDao.java
@@ -3,6 +3,6 @@ package com.pe.shopping.cart.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.pe.shopping.cart.entity.Venta;
 
-public interface VentaDao extends JpaRepository<Venta, Integer>{
+public interface VentaDao extends JpaRepository<Venta, Long>{
 
 }
